requirements.txt
mss~=10.1.0
pillow~=12.1.0
pynput~=1.8.1
opencv-python~=4.13.0.92
numpy~=2.4.2
reportlab~=4.4.9
python-docx~=1.2.0